refactor: remove ProvidesBindings trait
The reason is that Laravel provides this functionality out of the box
and actually it conflicts a bit. If you've noticed all the occurrences
of `$singletons = ...` are public declarations. This is because the
Laravel's check if the property exists but not that it is public.
However, it accesses it as if it was public.

I thought about leaving the trait there, but it relies on `$this` to
have an app-property which may not be the case for other non-provider
instances.